(improvement)(semantic) optimized llms2qlParser code and support prior exts to llm
