+++ chat/core/src/test/java/com/tencent/supersonic/chat/parser/llm/s2ql/LLMS2QLParserTest.java
@@ -4,8 +4,6 @@ import static org.mockito.Mockito.when;
 
 import com.tencent.supersonic.chat.api.pojo.SchemaElement;
 import com.tencent.supersonic.chat.api.pojo.SchemaValueMap;
-import com.tencent.supersonic.chat.api.pojo.SemanticCorrectInfo;
-import com.tencent.supersonic.chat.api.pojo.SemanticParseInfo;
 import com.tencent.supersonic.chat.api.pojo.SemanticSchema;
 import com.tencent.supersonic.common.util.ContextUtils;
 import com.tencent.supersonic.knowledge.service.SchemaService;
@@ -62,19 +60,5 @@ class LLMS2QLParserTest {
 
         when(mockSchemaService.getSemanticSchema()).thenReturn(mockSemanticSchema);
         mockContextUtils.when(() -> ContextUtils.getBean(SchemaService.class)).thenReturn(mockSchemaService);
-
-        SemanticParseInfo parseInfo = new SemanticParseInfo();
-        SchemaElement model = new SchemaElement();
-        model.setId(2L);
-        parseInfo.setModel(model);
-        SemanticCorrectInfo semanticCorrectInfo = SemanticCorrectInfo.builder()
-                .sql("select count(song_name) from 歌曲库 where singer_name = '周先生' and YEAR(publish_time) >= 2023 ")
-                .parseInfo(parseInfo)
-                .build();
-
-        LLMS2QLParser llms2QLParser = new LLMS2QLParser();
-
-        llms2QLParser.updateParseInfo(semanticCorrectInfo, 2L, parseInfo);
-
     }
 }
